Chapter04/cloud_function/code/requirements.txt
google-cloud-storage
functions-framework==3.*
